+++ lib/crucible/tensor_ref.ex
@@ -1,0 +1,10 @@
+defmodule Crucible.TensorRef do
+  @moduledoc """
+  V4 pointer to externally stored raw tensor bytes.
+  """
+
+  @derive Jason.Encoder
+  defstruct [:uri, :digest, :shape, :dtype, :byte_size, :format]
+
+  @type t :: %__MODULE__{}
+end